feat: split output.py into output/render.py and output/formatters.py
